Don't delete dist folder to preserve a potential symlink
Summary:
Ysul is configured (see D1391) to serve /var/51-wwwroot/rain
as https//rain51.nasqueron.org.

To serve there the development version, the dist/ folder must
be a symlink to this rain51 folder.

But currently, `npm start` deletes the symlink to recreate a 'dist' folder.
If gump deletes only the content of the dist, symlink is kept and
development version is still served.

Test Plan:
  - `npm start`
  - `file dist`
  - browse https://rain51.nasqueron.org

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D1392